refactor: move brc-20 into meta-protocols dir
