+++ jest.config.js
@@ -157,10 +157,10 @@ module.exports = {
     // testLocationInResults: false,
   
     // The glob patterns Jest uses to detect test files
-    // testMatch: [
-    //   "**/__tests__/**/*.[jt]s?(x)",
-    //   "**/?(*.)+(spec|test).[tj]s?(x)"
-    // ],
+    testMatch: [
+      "**/__tests__/**/?(*.)+(spec|test).[tj]s?(x)",
+      "**/?(*.)+(spec|test).[tj]s?(x)"
+    ],
   
     // An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
     testPathIgnorePatterns: [
