+++ src/meta-protocols/brc-20/pg/helpers.ts
@@ -1,8 +1,8 @@
 import { Static, Type } from '@fastify/type-provider-typebox';
 import { TypeCompiler } from '@sinclair/typebox/compiler';
 import BigNumber from 'bignumber.js';
-import { hexToBuffer } from '../../api/util/helpers';
-import { DbInscriptionInsert } from '../types';
+import { hexToBuffer } from '../../../api/util/helpers';
+import { DbInscriptionInsert } from '../../../pg/types';
 
 const Brc20TickerSchema = Type.String({ minLength: 1 });
 const Brc20NumberSchema = Type.RegEx(/^((\d+)|(\d*\.?\d+))$/);
